Got lots working (all of which needs documentation)
There is currently a weird (bug?) with omega being above or below 2000, due to multiplication. Not sure if it is even a bug, or just an effect or something tho. Needs more work in that.
I also wrote a follower with PIDF that should hopefully be easy to convert over to Robot code (pretty independent of everything too!)